// src/app/models/add-programs.model.ts
export class Programs {

    id?: any;
    title?: string;
    description?: string;
    published?: boolean;
    picture?: string;
    details?: string;
    vactitle?: string;
    vacdetails?: string;
    applylink?: string;
    duration?: string;
    startdate?: string;
    cost?: string;
    requirements?: string;

  }